add module.h
